+++ lib/walex/event/event.ex
@@ -1,18 +1,3 @@
-defmodule WalEx.Event.Source do
-  @moduledoc false
-
-  @derive Jason.Encoder
-  defstruct([:name, :version, :db, :schema, :table, :columns])
-
-  @type t :: %WalEx.Event.Source{
-          version: String.t(),
-          db: String.t(),
-          schema: String.t(),
-          table: String.t(),
-          columns: map()
-        }
-end
-
 defmodule WalEx.Event do
   @moduledoc """
   Event DSL and casting
@@ -43,8 +28,6 @@ defmodule WalEx.Event do
     app_name = Keyword.get(opts, :name)
 
     quote do
-      import WalEx.Event.Dsl
-
       def filter_events(txn) do
         Event.filter_and_cast(unquote(app_name), txn)
       end
@@ -52,92 +35,6 @@ defmodule WalEx.Event do
       def filter_events(txn, name, type, filters) do
         Event.filter_and_cast(unquote(app_name), txn, name, type, filters)
       end
-    end
-  end
-
-  defmodule Dsl do
-    defmacro process_events_async(events, functions) do
-      module = hd(__CALLER__.context_modules)
-
-      quote do
-        Enum.each(unquote(events), fn event ->
-          unquote(functions)
-          |> Enum.each(fn function ->
-            task_fn =
-              case function do
-                # If function is a tuple, treat it as {Module, function}
-                {mod, func} when is_atom(mod) and is_atom(func) ->
-                  fn -> apply(mod, func, [event]) end
-
-                # If function is an atom, treat it as a local function in the current module
-                # maybe don't allow atoms
-                func when is_atom(func) ->
-                  fn -> apply(unquote(module), func, [event]) end
-
-                _ ->
-                  raise ArgumentError, "Invalid function: #{inspect(function)}"
-              end
-
-            Task.start(task_fn)
-          end)
-        end)
-      end
-    end
-
-    defmacro on_event(:all, do_block) do
-      quote do
-        def process_all(txn) do
-          case filter_events(txn) do
-            filtered_events when is_list(filtered_events) and filtered_events != [] ->
-              unquote(do_block).(filtered_events)
-
-            _ ->
-              {:error, :no_events}
-          end
-        end
-      end
-    end
-
-    defmacro on_event(name, filters \\ %{}, functions \\ [], do_block) do
-      quote do
-        def process_all(txn) do
-          case filter_events(txn, unquote(name), unquote(nil), unquote(filters)) do
-            filtered_events when is_list(filtered_events) and filtered_events != [] ->
-              process_events_async(filtered_events, unquote(functions))
-              unquote(do_block).(filtered_events)
-
-            _ ->
-              {:error, :no_events}
-          end
-        end
-      end
-    end
-
-    defp process_event(name, type, filters, functions, do_block) do
-      quote do
-        def unquote(:"process_#{type}")(txn) do
-          case filter_events(txn, unquote(name), unquote(type), unquote(filters)) do
-            filtered_events when is_list(filtered_events) and filtered_events != [] ->
-              process_events_async(filtered_events, unquote(functions))
-              unquote(do_block).(filtered_events)
-
-            _ ->
-              {:error, :no_events}
-          end
-        end
-      end
-    end
-
-    defmacro on_insert(name, filters \\ %{}, functions \\ [], do_block) do
-      process_event(name, :insert, filters, functions, do_block)
-    end
-
-    defmacro on_update(name, filters \\ %{}, functions \\ [], do_block) do
-      process_event(name, :update, filters, functions, do_block)
-    end
-
-    defmacro on_delete(name, filters \\ %{}, functions \\ [], do_block) do
-      process_event(name, :delete, filters, functions, do_block)
     end
   end
